scanners/artifact-registry-scanning/src/vuln-to-gcs-cloud-function-init/cloud-function/requirements.txt
google-cloud-securitycenter
google-cloud-containeranalysis
google-cloud-storage